+++ src/lib/comms/sol-socket-impl.h
@@ -42,7 +42,7 @@ struct sol_socket_impl {
     int (*set_on_read)(struct sol_socket *s, bool (*cb)(void *data, struct sol_socket *s), const void *data);
     int (*set_on_write)(struct sol_socket *s, bool (*cb)(void *data, struct sol_socket *s), const void *data);
 
-    int (*recvmsg)(struct sol_socket *s, void *buf, size_t len, struct sol_network_link_addr *cliaddr);
+    ssize_t (*recvmsg)(struct sol_socket *s, void *buf, size_t len, struct sol_network_link_addr *cliaddr);
 
     int (*sendmsg)(struct sol_socket *s, const void *buf, size_t len,
         const struct sol_network_link_addr *cliaddr);
